enable caching of S3 etag request
decouple s3 service from verify service
