+++ src/main/java/edu/harvard/drs/verify/service/VerifyService.java
@@ -18,39 +18,19 @@ package edu.harvard.drs.verify.service;
 
 import static edu.harvard.drs.verify.utility.KeyUtility.buildKey;
 import static java.lang.String.format;
-import static org.apache.commons.lang3.StringUtils.removeEnd;
-import static org.apache.commons.lang3.StringUtils.removeStart;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import edu.harvard.drs.verify.config.AwsConfig;
 import edu.harvard.drs.verify.dto.OcflInventory;
 import edu.harvard.drs.verify.dto.VerificationError;
 import edu.harvard.drs.verify.exception.VerificationException;
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.URI;
 import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.context.annotation.RequestScope;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.awscore.exception.AwsServiceException;
-import software.amazon.awssdk.core.exception.SdkClientException;
-import software.amazon.awssdk.core.sync.ResponseTransformer;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.S3ClientBuilder;
-import software.amazon.awssdk.services.s3.model.GetObjectRequest;
-import software.amazon.awssdk.services.s3.model.HeadObjectRequest;
-import software.amazon.awssdk.services.s3.model.HeadObjectResponse;
-import software.amazon.awssdk.services.s3.model.InvalidObjectStateException;
-import software.amazon.awssdk.services.s3.model.NoSuchKeyException;
-import software.amazon.awssdk.services.s3.model.S3Exception;
 
 /**
  * Verify service.
@@ -60,34 +40,16 @@ import software.amazon.awssdk.services.s3.model.S3Exception;
 @RequestScope
 public class VerifyService {
 
-    private final String bucket;
-
-    private final S3Client s3;
-
-    private final ObjectMapper om;
+    private final S3Service s3Service;
 
     /**
-     * Construct verify service.
+     * Verify service constructor autowired.
+     *
+     * @param s3Service S3 service
      */
     @Autowired
-    public VerifyService(AwsConfig awsConfig) {
-        AwsBasicCredentials credentials = AwsBasicCredentials.create(
-            awsConfig.getAccessKeyId(),
-            awsConfig.getSecretAccessKey()
-        );
-
-        S3ClientBuilder builder = S3Client.builder()
-            .region(awsConfig.getRegion())
-            .credentialsProvider(StaticCredentialsProvider.create(credentials));
-
-        if (StringUtils.isNotEmpty(awsConfig.getEndpointOverride())) {
-            log.info("AWS endpoint override: {}", awsConfig.getEndpointOverride());
-            builder = builder.endpointOverride(URI.create(awsConfig.getEndpointOverride()));
-        }
-
-        this.bucket = awsConfig.getBucketName();
-        this.s3 = builder.build();
-        this.om = new ObjectMapper();
+    public VerifyService(S3Service s3Service) {
+        this.s3Service = s3Service;
     }
 
     /**
@@ -118,55 +80,13 @@ public class VerifyService {
         verify(id, input, true);
     }
 
-    /**
-     * Fetch OCFL inventoy.json from S3 and serialize.
-     *
-     * @param id DRS id
-     * @return serialize OCFL inventory
-     * @throws NoSuchKeyException the specified key does not exist
-     * @throws InvalidObjectStateException object is archived and inaccessible until restored
-     * @throws AwsServiceException something went wrong with S3 request
-     * @throws SdkClientException something went wrong with S3 request
-     * @throws S3Exception something went wrong with S3 request
-     * @throws IOException something went wrong serializing OCFL inventory
-     */
-    OcflInventory fetchInventory(Long id) throws NoSuchKeyException, InvalidObjectStateException,
-        AwsServiceException, SdkClientException, S3Exception, IOException {
-
-        String key = buildKey(id, "inventory.json");
-
-        GetObjectRequest request = GetObjectRequest.builder()
-            .bucket(bucket)
-            .key(key)
-            .build();
-
-        try (InputStream is = this.s3.getObject(request, ResponseTransformer.toInputStream())) {
-            return this.om.readValue(is, OcflInventory.class);
-        }
-    }
-
-    /**
-     * Request head object from S3 for given key.
-     *
-     * @param key S3 object key
-     * @return S3 head object response
-     */
-    HeadObjectResponse getHeadObject(String key) {
-        HeadObjectRequest request = HeadObjectRequest.builder()
-            .bucket(bucket)
-            .key(key)
-            .build();
-
-        return this.s3.headObject(request);
-    }
-
     private OcflInventory verify(Long id, Map<String, String> input) throws IOException, VerificationException {
         return verify(id, input, false);
     }
 
     private OcflInventory verify(Long id, Map<String, String> input, boolean update)
         throws IOException, VerificationException {
-        OcflInventory inventory = fetchInventory(id);
+        OcflInventory inventory = s3Service.fetchInventory(id);
 
         Map<String, VerificationError> errors = new ConcurrentHashMap<>();
 
@@ -182,9 +102,7 @@ public class VerifyService {
                     String expected = entry.getValue();
 
                     try {
-                        HeadObjectResponse response = getHeadObject(key);
-
-                        String actual = removeEnd(removeStart(response.eTag(), "\""), "\"");
+                        String actual = s3Service.getHeadObjectEtag(key);
 
                         if (!expected.equals(actual)) {
                             VerificationError error = VerificationError.builder()
